% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zomato-api.R
\name{zomato}
\alias{zomato}
\title{Zomato API}
\description{
An API key is needed to bridge communication between R and Zomato. You can
sign up the key \href{https://developers.zomato.com/api}{here}.
}
\section{Methods}{

\strong{Restaurants:}
\enumerate{
\item dailymenu(res_id)
\item restaurant(res_id)
\item reviews(res_id)
\item search(query, lat = NULL, lon = NULL, sort = NULL, order = NULL)
\strong{Location:}
\item location_details(query, lat = NULL, lon = NULL)
\item locations(entity_id, entity_type)
\strong{Common:}
\item categories()
\item cities(query, lat = NULL, lon = NULL, city_ids = NULL)
\item collections(city_id = NULL, lat = NULL, lon = NULL)
\item cuisines(city_id = NULL, lat = NULL, lon = NULL)
\item establishments(city_id = NULL, lat = NULL, lon = NULL)
\item geocode(lat, lon)
}
}

\examples{
zmt <- zomato$new("your-api-key")
zmt
bbb <- zmt$search(query = "Brother Budan Baba Melbourne")
zmt$reviews(res_id = bbb$id[1])
zmt$restaurant(res_id = bbb$id[1])
zmt$dailymenu(res_id = 16507624)

zmt$locations(query = "Melbourne")
zmt$locations(query = "Melbourne", -37.8136, 144.9631)
zmt$location_details(93747, "zone")

zmt$categories()
zmt$cities(query = "Melbourne")
zmt$collections(259)
zmt$cuisines(259)
zmt$establishments(259)
zmt$geocode(-37.8136, 144.9631)
}
