% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zomato-api.R
\name{zomato}
\alias{zomato}
\title{Query the Zomato API}
\description{
An API key is needed to bridge communication between R and Zomato. You can
sign up the key \href{https://developers.zomato.com/api}{here}.
}
\section{Methods}{

\strong{Restaurants:}
\itemize{
\item \code{dailymenu(res_id)}
\itemize{
\item Get daily menu using Zomato restaurant ID.
}
\item \code{restaurant(res_id)}
\itemize{
\item Get detailed restaurant information using Zomato restaurant ID.
}
\item \code{reviews(res_id)}
\itemize{
\item Get restaurant reviews using the Zomato restaurant ID. Only 5 latest
reviews are available under the Basic API plan.
}
\item \code{search( query = NULL, lat = NULL, lon = NULL, radius = NULL, cuisines = NULL, establishment_type = NULL, collection_id = NULL, category = NULL, sort = NULL, order = NULL )}
\itemize{
\item The location input can be specified using Zomato location ID or coordinates.
cuisine/establishment/ collection IDs can be obtained from respective
api calls.
}
}

\strong{Location:}
\itemize{
\item \code{location_details(entity_id, entity_type)}
\itemize{
\item Get Foodie Index, Nightlife Index, Top Cuisines and Best rated
restaurants in a given location
}
\item \code{locations(query, lat = NULL, lon = NULL)}
\itemize{
\item Search for Zomato locations by keyword. Provide coordinates to get better
search results.
}
}

\strong{Common:}
\itemize{
\item \code{categories()}
\itemize{
\item Get a list of categories. List of all restaurants categorized under a
particular restaurant type can be obtained using /search API with
}
\item \code{cities(query, lat = NULL, lon = NULL, city_ids = NULL)}
\itemize{
\item Find the Zomato ID and other details for a city.
}
\item \code{collections(city_id = NULL, lat = NULL, lon = NULL)}
\itemize{
\item Returns Zomato Restaurant Collections in a city.
}
\item \code{cuisines(city_id = NULL, lat = NULL, lon = NULL)}
\itemize{
\item Get a list of all cuisines of restaurants listed in a city.
}
\item \code{establishments(city_id = NULL, lat = NULL, lon = NULL)}
\itemize{
\item Get a list of restaurant types in a city.
}
\item \code{geocode(lat, lon)}
\itemize{
\item Get Foodie and Nightlife Index, list of popular cuisines and nearby
restaurants around the given coordinates.
}
}
}

\section{Arguments}{

\itemize{
\item \code{res_id}, \code{entity_id}, \code{city_id}: identifiers in integer
\item \code{query}: string that you search for
\item \code{lat}, \code{lon}: geo coordinates in double
\item \code{entity_type}: string
\item \code{sort}: \code{NULL}, "cost", "rating", "real_distance"
\item \code{order}: \code{NULL}, "asc", "desc"
\item Parameters without defaults are required to specify.
}
}

\examples{
\dontrun{
zmt <- zomato$new("your-api-key")
zmt
bbb <- zmt$search(query = "Brother Budan Baba Melbourne")
zmt$reviews(res_id = bbb$id[1])
zmt$restaurant(res_id = bbb$id[1])
zmt$dailymenu(res_id = 16507624)

zmt$locations(query = "Melbourne")
zmt$locations(query = "Melbourne", -37.8136, 144.9631)
zmt$location_details(93747, "zone")

zmt$categories()
zmt$cities(query = "Melbourne")
zmt$collections(259)
zmt$cuisines(259)
zmt$establishments(259)
zmt$geocode(-37.8136, 144.9631)
}
}
\references{
\link{https://developers.zomato.com/documentation}
}
